toml_utils: adsp_config: Moved generic parser functions to toml_utils
A set of functions which make it easier to parse toml files has been
separated into a new file.
